refactor: modularize unzip
